--- packages/app/lib/services/url_launcher/url_launcher.dart
@@ -1,6 +1,0 @@
-import 'package:breizh_blok_mobile/domain/entities/domain_exception/domain_exception.dart';
-import 'package:fpdart/fpdart.dart';
-
-abstract interface class UrlLauncher {
-  TaskEither<DomainException, bool> openUrl(Uri uri);
-}
