+++ packages/app/test/mocks.dart
@@ -13,7 +13,7 @@ import 'package:breizh_blok_mobile/data/repositories/downloaded_boulder_reposito
 import 'package:breizh_blok_mobile/data/repositories/municipality/municipality_repository.dart';
 import 'package:breizh_blok_mobile/domain/repositories/boulder_feedback_repository.dart';
 import 'package:breizh_blok_mobile/services/share_content/share_content_service.dart';
-import 'package:breizh_blok_mobile/services/url_launcher/url_launcher.dart';
+import 'package:breizh_blok_url_launcher/breizh_blok_url_launcher.dart';
 import 'package:dio/dio.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/src/foundation/diagnostics.dart';
@@ -23,7 +23,6 @@ import 'package:location/location.dart';
 import 'package:mapbox_maps_flutter/mapbox_maps_flutter.dart';
 import 'package:mixpanel_flutter/mixpanel_flutter.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 class MockApiBoulderFeedbackDataSource extends Mock
     implements ApiBoulderFeedbackDataSource {}
@@ -63,16 +62,6 @@ class MockBoulderFeedbackRepository extends Mock
     implements BoulderFeedbackRepository {}
 
 class MockImageBoulderCache extends Mock implements ImageBoulderCache {}
-
-class MockLaunchUrl extends Mock {
-  Future<bool> call(
-    Uri url, {
-    LaunchMode mode,
-    WebViewConfiguration webViewConfiguration,
-    BrowserConfiguration browserConfiguration,
-    String? webOnlyWindowName,
-  });
-}
 
 class MockLocation extends Mock implements Location {}
