+++ packages/url_launcher/analysis_options.yaml
@@ -1,0 +1,1 @@
+include: package:breizh_blok_lint/lint.yaml
