+++ packages/url_launcher/lib/src/exceptions.dart
@@ -1,0 +1,11 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'exceptions.freezed.dart';
+
+@freezed
+sealed class LaunchUrlException with _$LaunchUrlException implements Exception {
+  const LaunchUrlException._();
+
+  const factory LaunchUrlException.unknwown({required String message}) =
+      UnknownException;
+}
