+++ packages/url_launcher/lib/src/url_launcher.dart
@@ -1,0 +1,6 @@
+import 'package:breizh_blok_url_launcher/src/exceptions.dart';
+import 'package:fpdart/fpdart.dart';
+
+abstract interface class UrlLauncher {
+  TaskEither<LaunchUrlException, bool> openUrl(Uri uri);
+}
