+++ packages/url_launcher/pubspec.yaml
@@ -1,0 +1,25 @@
+name: breizh_blok_url_launcher
+publish_to: "none"
+version: 1.0.0
+
+resolution: workspace
+
+environment:
+  sdk: "^3.9.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+  fpdart: ^1.1.1
+  freezed_annotation: ^3.1.0
+  url_launcher: ^6.3.2
+
+dev_dependencies:
+  breizh_blok_lint:
+  build_runner: ^2.7.1
+  flutter_test:
+    sdk: flutter
+  freezed: ^3.2.3
+  mocktail: ^1.0.4
+
+flutter:
