+++ pubspec.yaml
@@ -11,6 +11,7 @@ workspace:
   - packages/app
   - packages/auth
   - packages/breizh_blok_api_generated
+  - packages/url_launcher
 
 melos:
   name: breizh_blok
