+++ sonar-project.properties
@@ -11,10 +11,13 @@ sonar.lang.patterns.dart=**/*.dart
 sonar.exclusions=**/*.g.dart,**/*.mocks.dart,**/app_localizations*.dart,**/*_generated/**
 sonar.dart.lcov.reportPaths=coverage/lcov.info
 
-sonar.modules=app,auth
+sonar.modules=app,auth,url_launcher
 
 app.sonar.projectBaseDir=packages/app
 app.sonar.flutter.tests.reportPath=sonar_tests.output
 
 auth.sonar.projectBaseDir=packages/auth
 auth.sonar.flutter.tests.reportPath=sonar_tests.output
+
+url_launcher.sonar.projectBaseDir=packages/url_launcher
+url_launcher.sonar.flutter.tests.reportPath=sonar_tests.output
